Add renderer flip options
