+++ packages/wasm-gerber-renderer/index.d.ts
@@ -34,6 +34,8 @@ export type FrameOptions = {
   background?: null | string | RGBAColor;
   fit?: boolean;
   padding?: number;
+  flipX?: boolean;
+  flipY?: boolean;
   view?: {
     zoomX: number;
     zoomY: number;
